Change CRA app to Vite app
For better perf and webpack support
