+++ src/components/ExampleArray.js
@@ -4,24 +4,24 @@ export const NFTS =
         "id": '#8764',
         "name": "nft1",
         "lodgeStatus": false,
-        "image": require('../assets/nft1.png')
+        "image": '../assets/nft1.png'
         },
         {
         "id": '#6764',
         "name": "nft2",
         "lodgeStatus": false,
-        "image": require('../assets/nft2.png')
+        "image": '../assets/nft2.png'
         },
         {
         "id": '#2347',
         "name": "nft3",
         "lodgeStatus": false,
-        "image": require('../assets/nft3.png')
+        "image": '../assets/nft3.png'
         },
         {
         "id": '#0987',
         "name": "nft4",
         "lodgeStatus": false,
-        "image": require('../assets/nft4.png')
+        "image": '../assets/nft4.png'
         }
     ]
