+++ src/components/TermsConditionsComponent.jsx
